+++ tests/examples/test_disease_model.py
@@ -29,7 +29,8 @@ def test_disease_model(fuzzy_checker: FuzzyChecker, disease_model_spec: Path) ->
 
     simulation = InteractiveContext(config)
 
-    pop = simulation.get_population()
+    all_attributes = simulation.get_attribute_names()
+    pop = simulation.get_population(all_attributes)
     expected_columns = {
         "is_alive",
         "previous_alive",