+++ tests/framework/test_engine.py
@@ -543,7 +543,7 @@ def test_private_columns_get_registered() -> None:
         "datetime_clock": [],
     }
     # Check that there are indeed other attributes registered besides via column_created
-    len(sim.get_population().columns) > 3
+    len(sim.get_population(sim.get_attribute_names()).columns) > 3
 
 
 ####################