+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+**4.0.1 - 04/02/26**
+
+- Remove support for calling InteractiveContext.get_population() without an explicit attribute request
+
 **4.0.0 - 04/02/26**
 -----------------------